+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.51.67](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.66...@standardnotes/mobile@3.51.67) (2023-03-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.51.66](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.65...@standardnotes/mobile@3.51.66) (2023-03-27)
 
 **Note:** Version bump only for package @standardnotes/mobile