+++ packages/web/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.150.34](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.33...@standardnotes/web@3.150.34) (2023-03-29)
+
+### Bug Fixes
+
+* Added backspace button on Super toolbar on Android as a workaround for backspace issue ([08a20e2](https://github.com/standardnotes/app/commit/08a20e25b416e672f0e80518ca66365f7089b4f3))
+
 ## [3.150.33](https://github.com/standardnotes/app/compare/@standardnotes/web@3.150.32...@standardnotes/web@3.150.33) (2023-03-27)
 
 **Note:** Version bump only for package @standardnotes/web